Final commit for sprint 2.
I updated the viewVitals view to include an animated gauge utilizing the gauge.js file.  I also updated the formatting, but it is far from finished.  Additionally, I moved some of the logic from the WebController to the Score.java file to clean up the controller.  An error is being reported from viewVitals, but it is a known issue and does not prevent the application from running.