Added comments for proposal.
I was able to figure out a good way to work with controllers and our project structure.  It is rudimentary and not secure by any means, but it is functional.  Proposed code has been commented out in each file in question.  The alternate controller are commented out at the bottom of the WebController.java file.